Forgot to commit in a while.
Moved to table format to display data.
Added bootstrap to make the display look more uniform to results on the web.
Added filter based on input field (from a 2-way bound angular term), to look specifically in
the `Facilities` property of the array passed.
Defaulted facilities to a warning message if no facilities are present in the data.
Added order by to Star Rating. Still needs ng-click function to turn on reverse sort.